make version overlay pixels not inconsistent

+++ fabric/src/main/java/net/justmili/leftforgotten/mixin/fabric/client/VersionOverlay.java
@@ -1,17 +1,18 @@
 package net.justmili.leftforgotten.mixin.fabric.client;
 
 import net.justmili.leftforgotten.init.LFResources;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Font;
-import net.minecraft.client.gui.Gui;
-import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.network.chat.Component;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.Gui;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.network.chat.Component;
 
 @Mixin(Gui.class)
 public abstract class VersionOverlay {
@@ -43,7 +44,7 @@ public abstract class VersionOverlay {
 
         guiGraphics.pose().pushPose();
         guiGraphics.pose().scale(1f / guiScaleFactor, 1f / guiScaleFactor, 1f);
-        guiGraphics.pose().scale(userScale, userScale, 1f);
+        guiGraphics.pose().scale((int) userScale, (int) userScale, 1f);
 
         guiGraphics.drawString(minecraft.font, component, drawX + 1, drawY + 1, textShadowColor, false);
         guiGraphics.drawString(minecraft.font, component, drawX, drawY, textColor, false);
